+++ playwright/helpers/run-helpers.ts
@@ -3,6 +3,7 @@ import { dataPath } from '../_project'
 import { AppConfig } from '../../src/common/app.d'
 import { GameState } from '../../src/common/game.d'
 import { DEFAULTS } from '../../src/common/defaults'
+import { SCORE_LIST_NAME } from '../../src/common/constants'
 
 export const testUrl = 'http://localhost:4173'
 
@@ -21,6 +22,12 @@ export const writeStorageState = async (page: Page): Promise<void> => {
   await page.context().storageState({ path: dataPath })
 }
 
+export const storeScoreData = async (page: Page, data: any): Promise<void> => {
+  await page.evaluate(data => {
+    localStorage.setItem(`mv-${SCORE_LIST_NAME}`, JSON.stringify(data))
+  }, data)
+}
+
 export const sleep = (duration: number): Promise<void> => {
     return new Promise((resolve) => {
         setTimeout(resolve, duration)
@@ -35,13 +42,16 @@ export const openPlayground = async (page: Page): Promise<void> => {
 
 export const visitAboutScreen = async (page: Page): Promise<void> => {
   await page.getByTitle('Description').click()
-  // when click has fully been processed, time can be measured
+  await sleep(10)
+}
+
+export const visitHallOfFameScreen = async (page: Page): Promise<void> => {
+  await page.getByTitle('Hall of Fame').click()
   await sleep(10)
 }
 
 export const visitConfigurationScreen = async (page: Page): Promise<void> => {
   await page.getByTitle('Settings').click()
-  // when click has fully been processed, time can be measured
   await sleep(10)
 }