+++ src/__mocks__/game-states.ts
@@ -153,7 +153,7 @@ export const wonGameState: GameState = {
   "score": {
     code: '3380IwBj5Yg', date: 1745518137216, user: 'Midas', rank: 1,
     game: {cells: 9, mines: 2, level: 8, effort: {least: 2, most: 7}},
-    signature: {fill_frequency: [18,17,6,4,2,0,0,0,0]},
+    signature: {fill_frequency: [18,17,6,4,2,0,0,0,0], invalid_code: false},
     play: {moves: 4, duration: 45.244},
     score: {efficiency: 0.5, speed: 0.08841, points: 44}
   },