+++ CHANGELOG.md
@@ -21,21 +21,31 @@
 - Calculate efficiency in sharp mode according to lower minimal number of moves
 - Similar Game button with score details, Soortgelijk Spel
 - Rename user with score details
-- Next/Previous button in ScorePopover
+- Count blanks and calculate blanks/pointers/mines ratio
 
 
 ## [4.1.0] - 2026
 
 ### Added
 - Histogram with fill distribution on Score popover
+- Browse scores in Score popover
+- Development tool for updating scorelist
+- Popover can be opened from diagram
+- Console warnings for invalid score codes
+- Next/Previous button in ScorePopover
+- Close button for scores-popover
+- Consent for deleting scores
 
 ### Changed
 - Boardsize parameter on configuration screen
 - Ease the transit from Intro to Game
 - More pace in pause animation
-
-### Fixed
-- Typing of scores
+- Refined score typing
+- Structured updater
+- Move delete and replay logic to popover 
+- Invalid codes logged with score.code
+- Removed scores with invalid codes
+- Keep list of scorelist-names to prevent reuse of names
 
 
 ## [4.0.0] - 2026-07-15